Turn on hibernate validation on app startup. Change some types to pass validation. Fixing validation problems with Postgres' GENERATED BY DEFAULT AS IDENTITY

+++ src/main/java/com/skovdev/springlearn/repository/jpa/UserSpringDataRepository.java
@@ -6,6 +6,6 @@ import com.skovdev.springlearn.model.User;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserSpringDataRepository extends EntityGraphJpaRepository<User, Long> {
+public interface UserSpringDataRepository extends EntityGraphJpaRepository<User, Integer> {
     User findByLogin(String login, EntityGraph entityGraph);
 }
